+++ tests/regression/register/tests.py
@@ -1,5 +1,7 @@
 from stdnet import test, orm
 
+from examples.models import SimpleModel
+
 
 class TestRegistration(test.TestCase):
 
@@ -13,3 +15,12 @@ class TestRegistration(test.TestCase):
         
     def testFlushModel(self):
         orm.flush_models()
+        
+    def testFlushExclude(self):
+        SimpleModel(code = 'test').save()
+        orm.flush_models(excludes = ('examples.simplemodel',))
+        qs = SimpleModel.objects.all()
+        self.assertTrue(qs)
+        orm.flush_models()
+        qs = SimpleModel.objects.all()
+        self.assertFalse(qs)